refactor: extract Topic Views from drawer
