+++ src/web/topic/components/Surface/TopicViews.styles.tsx
@@ -1,0 +1,6 @@
+import styled from "@emotion/styled";
+import { ListItemButton } from "@mui/material";
+
+export const NestedListItemButton = styled(ListItemButton)`
+  padding-left: 32px;
+`;
